rename module user
