+++ src/auctions/auctions.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common'
 import { PrismaService } from '../database/prisma.service'
-import { UsersService } from '../users/users.service'
+import { UsersService } from '../user/users.service'
 import { AuctionsResolver } from './auctions.resolver'
 import { AuctionsService } from './auctions.service'
