+++ src/auth/dto/sign.response.ts
@@ -1,5 +1,5 @@
 import { Field, ObjectType } from '@nestjs/graphql'
-import { User } from '../../users/entities/user.entity'
+import { User } from '../../user/entities/user.entity'
 
 @ObjectType()
 export class SignResponse {
